Template importer presents a list of templates
Improves the user experience by providing the user
a list of options instead of having to figure out
the semantic id of the template.

The list has a filter field to narrow the long list.

Also a select with the source of the template,
althought i doubt if this is something wanted.

The buttons below the form have been substituted
by a button at each element since the bottom buttons
supposed a click away of the center of the action now.

The button also gives feedback after being clicked,
and all the buttons get disabled.